@@ -9,7 +9,7 @@ namespace BitcoinBook.Test
 {
     public class PublicKeyTests
     {
-        readonly PublicKey key = new PublicKey(
+        readonly PublicKey key = new(
             "0887387e452b8eacc4acfde10d9aaf7f6d9a0f975aabb10d006e4da568744d06c",
             "061de6d95231cd89026e286df3b6ae4a894a3378e393e93a0f45b666329a0ae34");
 
