@@ -12,7 +12,7 @@ namespace BitcoinBook.Test
         const string baseUri = "http://song/api/";
         const string transactionId = "4ea6e2222c4d59dea646e21a103d8b812a6db433f8ca331778a9408990fa17ee";
 
-        readonly Mock<FakeHttpMessageHandler> mock = new Mock<FakeHttpMessageHandler> { CallBase = true };
+        readonly Mock<FakeHttpMessageHandler> mock = new() { CallBase = true };
         readonly TransactionFetcher fetcher;
 
         public TransactionFetcherTests()
@@ -41,7 +41,7 @@ namespace BitcoinBook.Test
             ExpectContent(transactionId, rawTransaction);
             var transaction = await fetcher.Fetch(transactionId);
             Assert.NotNull(transaction);
-            Assert.Equal(1, transaction.Version);
+            Assert.Equal(1, transaction!.Version);
             Assert.Single(transaction.Inputs);
             Assert.Equal("74e09a06fdf7dca73972629f129b51167483a040a3ac53bcdb1cd9a3a2e92abe", 
                 transaction.Inputs[0].PreviousTransaction.ToHex());
@@ -65,7 +65,8 @@ namespace BitcoinBook.Test
         public async Task FetchForRealTest()
         {
             var transaction = await IntegrationSetup.Mainnet.Fetcher.Fetch("0683c48ed57aad50b3c611366d522b11830c58f069de33bf5ceca7cafd44d98c");
-            Assert.Single(transaction.Inputs);
+            Assert.NotNull(transaction);
+            Assert.Single(transaction!.Inputs);
             Assert.Equal(2, transaction.Outputs.Count);
             Assert.Equal(0005897938, transaction.Outputs[0].Amount);
         }
