@@ -1,16 +1,13 @@
-﻿using System;
-using System.IO;
-
-namespace BitcoinBook
+﻿namespace BitcoinBook
 {
-    public class GetHeadersMessage : IMessage
+    public class GetHeadersMessage : MessageBase
     {
         public const int DefaultVersion = 70015;
 
         readonly byte[] startingBlock;
         readonly byte[] endingBlock;
 
-        public string Command => "getheaders";
+        public override string Command => "getheaders";
 
         public int Version { get; }
         public int Hashes { get; }
@@ -31,30 +28,19 @@ namespace BitcoinBook
 
         public static GetHeadersMessage Parse(byte[] bytes)
         {
-            var reader = new ByteReader(bytes);
-            try
-            {
-                return new GetHeadersMessage(
+            return Parse(bytes, reader => new GetHeadersMessage(
                     reader.ReadInt(4),
                     reader.ReadVarInt(),
                     reader.ReadBytes(32).ToReverseHex(),
-                    reader.ReadBytes(32).ToReverseHex());
-            }
-            catch (EndOfStreamException ex)
-            {
-                throw new FormatException("Read past end of data", ex);
-            }
+                    reader.ReadBytes(32).ToReverseHex()));
         }
 
-        public byte[] ToBytes()
+        public override void Write(ByteWriter writer)
         {
-            var stream = new MemoryStream();
-            var writer = new ByteWriter(stream);
             writer.Write(Version, 4);
             writer.WriteVar(Hashes);
             writer.Write(startingBlock);
             writer.Write(endingBlock);
-            return stream.ToArray();
         }
     }
 }
