@@ -5,8 +5,8 @@ namespace BitcoinBook
 {
     public interface ITransactionFetcher
     {
-        Task<Transaction> Fetch(string transactionId, bool fresh = false);
-        Task<Transaction> Fetch(byte[] transactionId, bool fresh = false);
+        Task<Transaction?> Fetch(string transactionId, bool fresh = false);
+        Task<Transaction?> Fetch(byte[] transactionId, bool fresh = false);
         Task<TransactionOutput[]> FetchPriorOutputs(IEnumerable<TransactionInput> inputs);
         Task<TransactionOutput> FetchPriorOutput(TransactionInput input);
         Task<TransactionOutput> FetchOutput(string outputPoint);
