@@ -10,16 +10,10 @@ namespace BitcoinBook
         public int Depth => GetDepth(Root);
         public int LeafCount => GetLeafCount(Root);
 
-        public MerkleTree(int leafCount)
-        {
-            if (leafCount < 1) throw new ArgumentException("leafCount must be > 0", nameof(leafCount));
-
-            Root = CreateTree(leafCount);
-        }
-
         public MerkleTree(IEnumerable<byte[]> hashes)
         {
-            var hashList = hashes?.ToList() ?? throw new ArgumentNullException(nameof(hashes));
+            if (hashes == null) throw new ArgumentNullException(nameof(hashes));
+            var hashList = hashes.ToList();
             CheckHashes(hashList, false);
 
             Root = CreateTree(hashList.Select(h => new MerkleNode(h)));
@@ -29,8 +23,7 @@ namespace BitcoinBook
         {
             if (proof == null) throw new ArgumentNullException(nameof(proof));
 
-            Root = CreateTree(proof.LeafCount);
-            Populate(Root, proof);
+            Root = CreateTree(proof, DepthNeeded(proof.LeafCount));
 
             CheckForLeftovers(proof.ProofHashes, nameof(proof.ProofHashes));
             CheckForLeftovers(proof.IncludedHashes, nameof(proof.IncludedHashes));
@@ -39,7 +32,8 @@ namespace BitcoinBook
 
         public MerkleProof CreateProof(IEnumerable<byte[]> includedHashes)
         {
-            var includedSet = new HashSet<string>(includedHashes?.Select(b => b.ToHex()) ?? throw new ArgumentNullException(nameof(includedHashes)));
+            if (includedHashes == null) throw new ArgumentNullException(nameof(includedHashes));
+            var includedSet = new HashSet<string>(includedHashes.Select(b => b.ToHex()));
             if (!includedSet.Any()) throw new ArgumentException("Included hashes must not be empty", nameof(includedHashes));
 
             var proof = new MerkleProof(LeafCount);
@@ -53,7 +47,7 @@ namespace BitcoinBook
             return proof;
         }
 
-        void AddToProof(MerkleNode node, MerkleProof proof, HashSet<string> includedSet)
+        void AddToProof(MerkleNode? node, MerkleProof proof, HashSet<string> includedSet)
         {
             if (node != null)
             {
@@ -75,35 +69,33 @@ namespace BitcoinBook
             }
         }
 
-        bool ContainsAny(MerkleNode node, HashSet<string> includedSet)
+        bool ContainsAny(MerkleNode? node, HashSet<string> includedSet)
         {
             return node != null && (includedSet.Contains(node.Hash.ToHex()) || ContainsAny(node.Left, includedSet) || ContainsAny(node.Right, includedSet));
         }
 
-        void Populate(MerkleNode node, MerkleProof proof)
+        MerkleNode CreateTree(MerkleProof proof, int depth)
         {
-            if (node != null)
+            if (proof.Flags.Dequeue())
             {
-                node.Hash = proof.Flags.Dequeue() ? proof.ProofHashes.Dequeue() :
-                    node.IsLeaf ?  proof.IncludedHashes.Dequeue() :
-                    ComputeHashFromChildren(node, proof);
+                return new(proof.ProofHashes.Dequeue());
             }
+
+            if (depth == 1)
+            {
+                return new(proof.IncludedHashes.Dequeue());
+            }
+
+            var left = CreateTree(proof, depth - 1);
+            var right = proof.Flags.Any() ? CreateTree(proof, depth - 1) : null;
+            return new(GetParentHash(left, right), left, right);
         }
 
-        byte[] ComputeHashFromChildren(MerkleNode node, MerkleProof proof)
-        {
-            Populate(node.Left, proof);
-            Populate(node.Right, proof);
-            return GetParentHash(node.Left, node.Right);
-        }
-
-        // ReSharper disable once ParameterOnlyUsedForPreconditionCheck.Local
         void CheckForLeftovers<T>(IEnumerable<T> items, string name)
         {
             if (items.Any()) throw new InvalidOperationException($"{name} has unused values");
         }
 
-        // ReSharper disable once ParameterOnlyUsedForPreconditionCheck.Local
         void CheckHashes(List<byte[]> hashes, bool isEmptyAllowed)
         {
             if (!(isEmptyAllowed || hashes.Any())) throw new ArgumentException("Must contain at least one hash", nameof(hashes));
@@ -111,30 +103,18 @@ namespace BitcoinBook
             if (!Unique(hashes)) throw new ArgumentException("All hashes must be unique", nameof(hashes));
         }
 
-        MerkleNode CreateTree(int leafCount)
-        {
-            var nodes = new List<MerkleNode>();
-            while (leafCount-- > 0)
-            {
-                nodes.Add(new MerkleNode());
-            }
-            return CreateTree(nodes);
-        }
-
         MerkleNode CreateTree(IEnumerable<MerkleNode> leafNodes)
         {
-            var nodes = leafNodes.ToList();
+            var nodes = new List<MerkleNode>(leafNodes);
 
             while (nodes.Count > 1)
             {
-                if (nodes.Count > 1 && nodes.Count % 2 == 1)
-                {
-                    nodes.Add(null);
-                }
                 var parents = new List<MerkleNode>();
                 for (var i = 0; i < nodes.Count; i += 2)
                 {
-                    parents.Add(new MerkleNode(GetParentHash(nodes[i], nodes[i + 1]), nodes[i], nodes[i + 1]));
+                    var left = nodes[i];
+                    var right = i + 1 < nodes.Count ? nodes[i + 1] : null;
+                    parents.Add(new MerkleNode(GetParentHash(left, right), left, right));
                 }
 
                 nodes = parents;
@@ -143,17 +123,17 @@ namespace BitcoinBook
             return nodes[0];
         }
 
-        byte[] GetParentHash(MerkleNode left, MerkleNode right)
+        byte[] GetParentHash(MerkleNode left, MerkleNode? right)
         {
-            return left.Hash == null ? null : Merkler.ComputeParent(left.Hash, right?.Hash ?? left.Hash);
+            return Merkler.ComputeParent(left.Hash, right?.Hash ?? left.Hash);
         }
 
-        int GetDepth(MerkleNode node)
+        int GetDepth(MerkleNode? node)
         {
             return node == null ? 0 : 1 + GetDepth(node.Left);
         }
 
-        int GetLeafCount(MerkleNode node)
+        int GetLeafCount(MerkleNode? node)
         {
             if (node == null)
             {
@@ -167,6 +147,19 @@ namespace BitcoinBook
         {
             var hashSet = new HashSet<string>();
             return valueList.Select(b => b.ToHex()).All(hashSet.Add);
+        }
+
+        int DepthNeeded(int leafCount)
+        {
+            var leaves = 1;
+            var depth = 1;
+            while (leafCount > leaves)
+            {
+                leaves *= 2;
+                depth += 1;
+            }
+
+            return depth;
         }
     }
 }
