@@ -1,7 +1,4 @@
-﻿using System;
-using System.IO;
-
-namespace BitcoinBook
+﻿namespace BitcoinBook
 {
     public class PingMessage : PingPongMessageBase
     {
@@ -13,15 +10,7 @@ namespace BitcoinBook
 
         public static PingMessage Parse(byte[] bytes)
         {
-            var reader = new ByteReader(bytes);
-            try
-            {
-                return new PingMessage(reader.ReadUnsignedLong(8));
-            }
-            catch (EndOfStreamException ex)
-            {
-                throw new FormatException("Read past end of data", ex);
-            }
+            return Parse(bytes, r => new PingMessage(r.ReadUnsignedLong(8)));
         }
     }
 }
