@@ -1,24 +1,17 @@
-﻿using System.IO;
-
-namespace BitcoinBook
+﻿namespace BitcoinBook
 {
-    public abstract class PingPongMessageBase : IMessage
+    public abstract class PingPongMessageBase : MessageBase
     {
         public ulong Nonce { get; }
-
-        public abstract string Command { get; }
 
         protected PingPongMessageBase(ulong nonce)
         {
             Nonce = nonce;
         }
 
-        public byte[] ToBytes()
+        public override void Write(ByteWriter writer)
         {
-            var stream = new MemoryStream();
-            var writer = new ByteWriter(stream);
             writer.Write(Nonce, 8);
-            return stream.ToArray();
         }
     }
 }
